Refactor HelpMenu. Add new params.
