+++ src/components/Help/HelpMenu.tsx
@@ -2,22 +2,17 @@ import * as React from "react";
 import { useContext, useMemo } from "react";
 import { Menu, Dropdown, Button } from "antd";
 import { QuestionCircleOutlined } from "@ant-design/icons";
-import ConfigContext from "../../../ConfigProvider/context";
-import "./index.less";
-
-export interface IHelpMenu {
-  onTutorialClick: () => void;
-  onHelperClick: () => void;
-  onVisibleChange: (visible: boolean) => void;
-  tourMenu?: React.ReactElement;
-  visible?: boolean;
-}
+import ConfigContext from "../ConfigProvider/context";
+import { IHelpMenu } from "./types";
 
 const HelpMenu: React.FC<IHelpMenu> = ({
-  onTutorialClick,
+  onTutorialLinkClick,
   onHelperClick,
   onVisibleChange,
   tourMenu,
+  tutorialDisabled,
+  helperDisabled,
+  btnDisabled,
   visible
 }) => {
   const { translate } = useContext(ConfigContext);
@@ -25,26 +20,37 @@ const HelpMenu: React.FC<IHelpMenu> = ({
   const menu = useMemo(() => {
     return (
       <Menu theme="light" className="help-menu-dropdown">
-        {onTutorialClick && (
-          <Menu.Item key="1" onClick={onTutorialClick}>
+        {onTutorialLinkClick && (
+          <Menu.Item
+            key="1"
+            disabled={tutorialDisabled}
+            onClick={onTutorialLinkClick}
+          >
             {translate("READ_TUTORIAL")}
           </Menu.Item>
         )}
         {onHelperClick && (
-          <Menu.Item key="2" onClick={onHelperClick}>
+          <Menu.Item key="2" disabled={helperDisabled} onClick={onHelperClick}>
             {translate("BES_HELPER")}
           </Menu.Item>
         )}
         {tourMenu && tourMenu}
       </Menu>
     );
-  }, [onTutorialClick, onHelperClick, translate, tourMenu]);
+  }, [
+    onTutorialLinkClick,
+    tutorialDisabled,
+    translate,
+    onHelperClick,
+    helperDisabled,
+    tourMenu
+  ]);
 
   const visibleProps = typeof visible === "boolean" ? { visible } : {};
 
   return (
     <>
-      {(onTutorialClick || onHelperClick || tourMenu) && (
+      {(onTutorialLinkClick || onHelperClick || tourMenu) && (
         <Dropdown
           {...visibleProps}
           onVisibleChange={onVisibleChange}
@@ -57,6 +63,7 @@ const HelpMenu: React.FC<IHelpMenu> = ({
             className="help-icon teaching-btn"
             onClick={e => e.preventDefault()}
             icon={<QuestionCircleOutlined />}
+            disabled={btnDisabled}
           ></Button>
         </Dropdown>
       )}
